+++ es/index.js
@@ -1,13 +1,14 @@
-import '@codetanzania/ewea-internals';
-import { pkg } from '@lykmapipo/common';
+import { PREDEFINE_NAMESPACE_ADMINISTRATIVELEVEL, PREDEFINE_NAMESPACE_ADMINISTRATIVEAREA, PREDEFINE_NAMESPACE_PARTYGROUP, PREDEFINE_NAMESPACE_PARTYROLE } from '@codetanzania/ewea-internals';
+import { safeMergeObjects, pkg } from '@lykmapipo/common';
 import { getString, apiVersion as apiVersion$1 } from '@lykmapipo/env';
 import { connect } from '@lykmapipo/mongoose-common';
 import { mount } from '@lykmapipo/express-common';
 import { Router, start as start$1 } from '@lykmapipo/express-rest-actions';
 import { createModels } from '@lykmapipo/file';
-import { isFunction, merge } from 'lodash';
+import { isFunction } from 'lodash';
 import { waterfall } from 'async';
 import { Party } from '@codetanzania/emis-stakeholder';
+import { DEFAULT_PREDEFINE_RELATION } from '@codetanzania/ewea-common';
 import { EventChangeLog, Event } from '@codetanzania/ewea-event';
 import { Predefine } from '@lykmapipo/predefine';
 
@@ -15,6 +16,18 @@ import { Predefine } from '@lykmapipo/predefine';
 // order: base to specific
 
 const PARTY_TYPE_FOCAL = 'Focal';
+const DEFAULT_RELATION_LEVEL = safeMergeObjects(DEFAULT_PREDEFINE_RELATION, {
+  namespace: PREDEFINE_NAMESPACE_ADMINISTRATIVELEVEL,
+});
+const DEFAULT_RELATION_AREA = safeMergeObjects(DEFAULT_PREDEFINE_RELATION, {
+  namespace: PREDEFINE_NAMESPACE_ADMINISTRATIVEAREA,
+});
+const DEFAULT_RELATION_GROUP = safeMergeObjects(DEFAULT_PREDEFINE_RELATION, {
+  namespace: PREDEFINE_NAMESPACE_PARTYGROUP,
+});
+const DEFAULT_RELATION_ROLE = safeMergeObjects(DEFAULT_PREDEFINE_RELATION, {
+  namespace: PREDEFINE_NAMESPACE_PARTYROLE,
+});
 
 // start: extra metric fields
 // order: base to specific
@@ -144,6 +157,30 @@ const PARTY_BASE_ROLE_PROJECTION = {
 
 /**
  * @constant
+ * @name PARTY_DEFAULT_PROJECTION
+ * @description Normalize Party fields with defaults before pass to
+ * the next stage in the pipeline.
+ * @type {object}
+ *
+ * @author lally elias <lallyelias87@gmail.com>
+ * @license MIT
+ * @since 0.3.0
+ * @version 0.1.0
+ */
+const PARTY_DEFAULT_PROJECTION = {
+  _id: 1,
+  type: 1,
+  createdAt: 1,
+  updatedAt: 1,
+  metrics: 1,
+  level: { $ifNull: ['$level', DEFAULT_RELATION_LEVEL] },
+  area: { $ifNull: ['$area', DEFAULT_RELATION_AREA] },
+  group: { $ifNull: ['$group', DEFAULT_RELATION_GROUP] },
+  role: { $ifNull: ['$role', DEFAULT_RELATION_ROLE] },
+};
+
+/**
+ * @constant
  * @name PARTY_BASE_PROJECTION
  * @description Party fields passed to the next stage in the pipeline.
  * @type {object}
@@ -171,7 +208,7 @@ const PARTY_BASE_PROJECTION = {
 /**
  * @constant
  * @name PARTY_FACET_OVERVIEW
- * @description Party overview facet.
+ * @description General `Party` overview facet.
  * @type {object}
  *
  * @author lally elias <lallyelias87@gmail.com>
@@ -198,6 +235,150 @@ const PARTY_FACET_OVERVIEW = {
     {
       $project: {
         _id: 0,
+      },
+    },
+  ],
+};
+
+/**
+ * @constant
+ * @name PARTY_FACET_OVERALL_LEVEL
+ * @description Overall `Party` breakdown by administrative level facet.
+ * @type {object}
+ *
+ * @author lally elias <lallyelias87@gmail.com>
+ * @license MIT
+ * @since 0.3.0
+ * @version 0.1.0
+ */
+const PARTY_FACET_OVERALL_LEVEL = {
+  levels: [
+    {
+      $group: {
+        _id: '$level._id',
+        total: { $sum: 1 },
+        agency: { $sum: '$metrics.agency' },
+        focal: { $sum: '$metrics.focal' },
+        active: { $sum: '$metrics.active' },
+        inactive: { $sum: '$metrics.inactive' },
+        namespace: { $first: '$level.namespace' },
+        name: { $first: '$level.name' },
+        abbreviation: { $first: '$level.abbreviation' },
+        color: { $first: '$level.color' },
+        weight: { $first: '$level.weight' },
+      },
+    },
+    {
+      $sort: {
+        weight: 1,
+      },
+    },
+  ],
+};
+
+/**
+ * @constant
+ * @name PARTY_FACET_OVERALL_AREA
+ * @description Overall `Party` breakdown by administrative area facet.
+ * @type {object}
+ *
+ * @author lally elias <lallyelias87@gmail.com>
+ * @license MIT
+ * @since 0.3.0
+ * @version 0.1.0
+ */
+const PARTY_FACET_OVERALL_AREA = {
+  areas: [
+    {
+      $group: {
+        _id: '$area._id',
+        total: { $sum: 1 },
+        agency: { $sum: '$metrics.agency' },
+        focal: { $sum: '$metrics.focal' },
+        active: { $sum: '$metrics.active' },
+        inactive: { $sum: '$metrics.inactive' },
+        namespace: { $first: '$area.namespace' },
+        name: { $first: '$area.name' },
+        abbreviation: { $first: '$area.abbreviation' },
+        color: { $first: '$area.color' },
+        weight: { $first: '$area.weight' },
+      },
+    },
+    {
+      $sort: {
+        weight: 1,
+      },
+    },
+  ],
+};
+
+/**
+ * @constant
+ * @name PARTY_FACET_OVERALL_GROUP
+ * @description Overall `Party` breakdown by group facet.
+ * @type {object}
+ *
+ * @author lally elias <lallyelias87@gmail.com>
+ * @license MIT
+ * @since 0.3.0
+ * @version 0.1.0
+ */
+const PARTY_FACET_OVERALL_GROUP = {
+  groups: [
+    {
+      $group: {
+        _id: '$group._id',
+        total: { $sum: 1 },
+        agency: { $sum: '$metrics.agency' },
+        focal: { $sum: '$metrics.focal' },
+        active: { $sum: '$metrics.active' },
+        inactive: { $sum: '$metrics.inactive' },
+        namespace: { $first: '$group.namespace' },
+        name: { $first: '$group.name' },
+        abbreviation: { $first: '$group.abbreviation' },
+        color: { $first: '$group.color' },
+        weight: { $first: '$group.weight' },
+      },
+    },
+    {
+      $sort: {
+        weight: 1,
+      },
+    },
+  ],
+};
+
+/**
+ * @constant
+ * @name PARTY_FACET_OVERALL_ROLE
+ * @description Overall `Party` breakdown by role facet.
+ * @type {object}
+ *
+ * @author lally elias <lallyelias87@gmail.com>
+ * @license MIT
+ * @since 0.3.0
+ * @version 0.1.0
+ */
+const PARTY_FACET_OVERALL_ROLE = {
+  roles: [
+    {
+      $group: {
+        _id: '$role._id',
+        total: { $sum: 1 },
+        agency: { $sum: '$metrics.agency' },
+        focal: { $sum: '$metrics.focal' },
+        active: { $sum: '$metrics.active' },
+        inactive: { $sum: '$metrics.inactive' },
+        namespace: { $first: '$role.namespace' },
+        name: { $first: '$role.name' },
+        abbreviation: { $first: '$role.abbreviation' },
+        color: { $first: '$role.color' },
+        weight: { $first: '$role.weight' },
+      },
+    },
+    {
+      $sort: {
+        weight: 1,
       },
     },
   ],
@@ -231,6 +412,11 @@ const getPartyBaseAggregation = (criteria = {}) => {
   // initialize party base aggregation
   const base = Party.lookup(criteria);
 
+  // project default on relations
+  base.project(PARTY_DEFAULT_PROJECTION);
+
+  // TODO: project per relations before add metrics
+
   // add extra metric fields
   base.addFields(PARTY_BASE_METRIC_FIELDS);
 
@@ -240,6 +426,8 @@ const getPartyBaseAggregation = (criteria = {}) => {
   // return party base aggregation
   return base;
 };
+
+// TODO: metrics only aggregation; getPartyMetrics/getPartyOverview
 
 /**
  * @function getPartyAnalysis
@@ -252,7 +440,7 @@ const getPartyBaseAggregation = (criteria = {}) => {
  * @author lally elias <lallyelias87@gmail.com>
  * @license MIT
  * @since 0.2.0
- * @version 0.1.0
+ * @version 0.2.0
  * @static
  * @public
  * @example
@@ -270,21 +458,41 @@ const getPartyAnalysis = (criteria, done) => {
   const base = getPartyBaseAggregation(filter);
 
   // add facets
-  base.facet(PARTY_FACET_OVERVIEW);
+  const facets = {
+    ...PARTY_FACET_OVERVIEW,
+    ...PARTY_FACET_OVERALL_LEVEL,
+    ...PARTY_FACET_OVERALL_AREA,
+    ...PARTY_FACET_OVERALL_GROUP,
+    ...PARTY_FACET_OVERALL_ROLE,
+  };
+  base.facet(facets);
 
   // run aggregation
   const aggregate = (next) => base.exec(next);
   const normalize = (result, next) => {
     // TODO: extract to utils
-    const data = merge(...result);
-    data.overview = merge(...data.overview);
+
+    // ensure data
+    const { overview, levels, areas, groups, roles } = safeMergeObjects(
+      ...result
+    );
+
+    // normalize result
+    const data = safeMergeObjects({
+      overview: safeMergeObjects(...overview),
+      overall: { levels, areas, groups, roles },
+    });
+
+    // return normalize result
     return next(null, { data });
   };
-  const tasks = [aggregate, normalize];
 
   // return
+  const tasks = [aggregate, normalize];
   return waterfall(tasks, cb);
 };
+
+// TODO: dataset only aggregation; getPartyDatasets
 
 /* constants */
 const API_VERSION = getString('API_VERSION', '1.0.0');
@@ -706,4 +914,4 @@ const start = (done) => {
   });
 };
 
-export { CHANGELOG_FLAG_METRIC_FIELDS, CHANGELOG_TIME_METRIC_FIELDS, EVENT_FLAG_METRIC_FIELDS, EVENT_TIME_METRIC_FIELDS, PARTY_BASE_AREA_PROJECTION, PARTY_BASE_GROUP_PROJECTION, PARTY_BASE_LEVEL_PROJECTION, PARTY_BASE_METRIC_FIELDS, PARTY_BASE_PROJECTION, PARTY_BASE_ROLE_PROJECTION, PARTY_FACET_OVERVIEW, PARTY_TYPE_FOCAL, PREDEFINE_FLAG_METRIC_FIELDS, PREDEFINE_TIME_METRIC_FIELDS, apiVersion, getChangeLogBaseAggregation, getEventBaseAggregation, getPartyAnalysis, getPartyBaseAggregation, getPredefineBaseAggregation, info, router as reportRouter, start };
+export { CHANGELOG_FLAG_METRIC_FIELDS, CHANGELOG_TIME_METRIC_FIELDS, EVENT_FLAG_METRIC_FIELDS, EVENT_TIME_METRIC_FIELDS, PARTY_BASE_AREA_PROJECTION, PARTY_BASE_GROUP_PROJECTION, PARTY_BASE_LEVEL_PROJECTION, PARTY_BASE_METRIC_FIELDS, PARTY_BASE_PROJECTION, PARTY_BASE_ROLE_PROJECTION, PARTY_DEFAULT_PROJECTION, PARTY_FACET_OVERALL_AREA, PARTY_FACET_OVERALL_GROUP, PARTY_FACET_OVERALL_LEVEL, PARTY_FACET_OVERALL_ROLE, PARTY_FACET_OVERVIEW, PREDEFINE_FLAG_METRIC_FIELDS, PREDEFINE_TIME_METRIC_FIELDS, apiVersion, getChangeLogBaseAggregation, getEventBaseAggregation, getPartyAnalysis, getPartyBaseAggregation, getPredefineBaseAggregation, info, router as reportRouter, start };